refactor: Update library files in correct way
